+++ RELEASENOTES.md
@@ -306,6 +306,7 @@
         use in `DefaultCastOptionsProvider#toMediaQueueItem`, when available.
     *   Enable remote to local transfers in `DefaultCastOptionsProvider`.
     *   Add support for Cast in the Session demo.
+    *   Add support for displaying a media route button on a Composable UI.
 *   Test Utilities:
     *   Add maximum time diff for the auto-advancing behavior of `FakeClock`. It
         defaults to 1 second, but is configurable via `FakeClock.Builder`.